+++ test/dynamics/models/test_hamiltonian_model.py
@@ -11,15 +11,17 @@
 # that they have been altered from the originals.
 # pylint: disable=invalid-name
 
-"""tests for quantum_models.HamiltonianModel"""
+"""tests for qiskit_dynamics.models.HamiltonianModel"""
 
 import numpy as np
+from jax import jit, grad
 from scipy.linalg import expm
 from qiskit.quantum_info.operators import Operator
 from qiskit_dynamics.models import HamiltonianModel
 from qiskit_dynamics.signals import Signal, SignalList
 from qiskit_dynamics.dispatch import Array
 from ..common import QiskitDynamicsTestCase, TestJaxBase
+from .test_operator_collections import _wrap
 
 
 class TestHamiltonianModel(QiskitDynamicsTestCase):
@@ -40,7 +42,7 @@ class TestHamiltonianModel(QiskitDynamicsTestCase):
         self.r = r
         self.basic_hamiltonian = HamiltonianModel(operators=operators, signals=signals)
 
-    def _basic_frame_evaluate_test(self, frame_operator, t):
+    def _basic_frame_evaluate_generator_test(self, frame_operator, t):
         """Routine for testing setting of valid frame operators using
         basic_hamiltonian.
 
@@ -54,7 +56,7 @@ class TestHamiltonianModel(QiskitDynamicsTestCase):
             t (float): time of frame transformation
         """
 
-        self.basic_hamiltonian.frame = frame_operator
+        self.basic_hamiltonian.rotating_frame = frame_operator
 
         # convert to 2d array
         if isinstance(frame_operator, Operator):
@@ -62,7 +64,7 @@ class TestHamiltonianModel(QiskitDynamicsTestCase):
         if isinstance(frame_operator, Array) and frame_operator.ndim == 1:
             frame_operator = np.diag(frame_operator)
 
-        value = self.basic_hamiltonian.evaluate(t)
+        value = self.basic_hamiltonian(t) / -1j
 
         twopi = 2 * np.pi
 
@@ -86,41 +88,41 @@ class TestHamiltonianModel(QiskitDynamicsTestCase):
         set up basic hamiltonian.
         """
 
-        self._basic_frame_evaluate_test(Array([1.0, -1.0]), 1.123)
-        self._basic_frame_evaluate_test(Array([1.0, -1.0]), np.pi)
+        self._basic_frame_evaluate_generator_test(Array([1.0, -1.0]), 1.123)
+        self._basic_frame_evaluate_generator_test(Array([1.0, -1.0]), np.pi)
 
     def test_non_diag_frame_operator_basic_hamiltonian(self):
         """Test setting a non-diagonal frame operator for the internally
         set up basic model.
         """
-        self._basic_frame_evaluate_test(self.Y + self.Z, 1.123)
-        self._basic_frame_evaluate_test(self.Y - self.Z, np.pi)
+        self._basic_frame_evaluate_generator_test(self.Y + self.Z, 1.123)
+        self._basic_frame_evaluate_generator_test(self.Y - self.Z, np.pi)
 
-    def test_evaluate_no_frame_basic_hamiltonian(self):
-        """Test evaluation without a frame in the basic model."""
+    def test_evaluate_generator_no_frame_basic_hamiltonian(self):
+        """Test generator evaluation without a frame in the basic model."""
 
         t = 3.21412
-        value = self.basic_hamiltonian.evaluate(t)
+        value = self.basic_hamiltonian(t) / -1j
         twopi = 2 * np.pi
         d_coeff = self.r * np.cos(2 * np.pi * self.w * t)
         expected = twopi * self.w * self.Z.data / 2 + twopi * d_coeff * self.X.data / 2
 
         self.assertAllClose(value, expected)
 
-    def test_evaluate_in_frame_basis_basic_hamiltonian(self):
-        """Test evaluation in frame basis in the basic_model."""
+    def test_evaluate_generator_in_frame_basis_basic_hamiltonian(self):
+        """Test generator evaluation in frame basis in the basic_hamiltonian."""
 
         frame_op = (self.X + 0.2 * self.Y + 0.1 * self.Z).data
 
         # enter the frame given by the -1j * X
-        self.basic_hamiltonian.frame = frame_op
+        self.basic_hamiltonian.rotating_frame = frame_op
 
-        # get the frame basis used in model. Note that the RotatingFrame object
+        # get the frame basis used in model. Note that the Frame object
         # orders the basis according to the ordering of eigh
         _, U = np.linalg.eigh(frame_op)
 
         t = 3.21412
-        value = self.basic_hamiltonian.evaluate(t, in_frame_basis=True)
+        value = self.basic_hamiltonian(t, in_frame_basis=True) / -1j
 
         # compose the frame basis transformation with the exponential
         # frame rotation (this will be multiplied on the right)
@@ -137,7 +139,7 @@ class TestHamiltonianModel(QiskitDynamicsTestCase):
 
         self.assertAllClose(value, expected)
 
-    def test_evaluate_pseudorandom(self):
+    def test_evaluate_generator_pseudorandom(self):
         """Test evaluate with pseudorandom inputs."""
 
         rng = np.random.default_rng(30493)
@@ -163,7 +165,9 @@ class TestHamiltonianModel(QiskitDynamicsTestCase):
         rand_carriers = Array(rng.uniform(low=-b, high=b, size=(num_terms)))
         rand_phases = Array(rng.uniform(low=-b, high=b, size=(num_terms)))
 
-        self._test_evaluate(frame_op, randoperators, rand_coeffs, rand_carriers, rand_phases)
+        self._test_evaluate_generator(
+            frame_op, randoperators, rand_coeffs, rand_carriers, rand_phases
+        )
 
         rng = np.random.default_rng(94818)
         num_terms = 5
@@ -188,9 +192,11 @@ class TestHamiltonianModel(QiskitDynamicsTestCase):
         rand_carriers = Array(rng.uniform(low=-b, high=b, size=(num_terms)))
         rand_phases = Array(rng.uniform(low=-b, high=b, size=(num_terms)))
 
-        self._test_evaluate(frame_op, randoperators, rand_coeffs, rand_carriers, rand_phases)
+        self._test_evaluate_generator(
+            frame_op, randoperators, rand_coeffs, rand_carriers, rand_phases
+        )
 
-    def _test_evaluate(self, frame_op, operators, coefficients, carriers, phases):
+    def _test_evaluate_generator(self, frame_op, operators, coefficients, carriers, phases):
 
         sig_list = []
         for coeff, freq, phase in zip(coefficients, carriers, phases):
@@ -204,9 +210,9 @@ class TestHamiltonianModel(QiskitDynamicsTestCase):
 
             sig_list.append(Signal(get_env_func(), freq, phase))
         sig_list = SignalList(sig_list)
-        model = HamiltonianModel(operators, sig_list, frame=frame_op)
+        model = HamiltonianModel(operators, drift=None, signals=sig_list, rotating_frame=frame_op)
 
-        value = model.evaluate(1.0)
+        value = model(1.0, in_frame_basis=False) / -1j
         coeffs = np.real(coefficients * np.exp(1j * 2 * np.pi * carriers * 1.0 + 1j * phases))
         expected = (
             expm(1j * np.array(frame_op))
@@ -214,38 +220,10 @@ class TestHamiltonianModel(QiskitDynamicsTestCase):
             @ expm(-1j * np.array(frame_op))
             - frame_op
         )
+        self.assertAllClose(model._signals(1), coeffs)
+        self.assertAllClose(model.get_operators(), operators)
 
         self.assertAllClose(value, expected)
-
-    def test_cutoff_freq(self):
-        """Test evaluation with a cutoff frequency."""
-
-        # enter frame of drift
-        self.basic_hamiltonian.frame = self.basic_hamiltonian.drift
-
-        # set cutoff freq to 2 * drive freq (standard RWA)
-        self.basic_hamiltonian.cutoff_freq = 2 * self.w
-
-        # result should just be the X term halved
-        eval_rwa = self.basic_hamiltonian.evaluate(2.0)
-        expected = 2 * np.pi * (self.r / 2) * self.X.data / 2
-        self.assertAllClose(eval_rwa, expected)
-
-        def drive_func(t):
-            return t ** 2 + t ** 3 * 1j
-
-        self.basic_hamiltonian.signals = [self.w, Signal(drive_func, self.w)]
-
-        # result should now contain both X and Y terms halved
-        t = 2.1231 * np.pi
-        dRe = np.real(drive_func(t))
-        dIm = np.imag(drive_func(t))
-        eval_rwa = self.basic_hamiltonian.evaluate(t)
-        expected = (
-            2 * np.pi * (self.r / 2) * dRe * self.X.data / 2
-            + 2 * np.pi * (self.r / 2) * dIm * self.Y.data / 2
-        )
-        self.assertAllClose(eval_rwa, expected)
 
 
 class TestHamiltonianModelJax(TestHamiltonianModel, TestJaxBase):
@@ -253,3 +231,30 @@ class TestHamiltonianModelJax(TestHamiltonianModel, TestJaxBase):
 
     Note: This class has no body but contains tests due to inheritance.
     """
+    def test_jitable_funcs(self):
+        """Tests whether all functions are jitable.
+        Checks if having a frame makes a difference, as well as 
+        all jax-compatible evaluation_modes."""
+        _wrap(jit,self.basic_hamiltonian.evaluate_generator)(1)
+        _wrap(jit,self.basic_hamiltonian.evaluate_rhs)(1,Array(np.array([0.2,0.4])))
+
+        self.basic_hamiltonian.rotating_frame = Array(np.array([[3j, 2j], [2j, 0]]))
+
+        _wrap(jit,self.basic_hamiltonian.evaluate_generator)(1)
+        _wrap(jit,self.basic_hamiltonian.evaluate_rhs)(1,Array(np.array([0.2,0.4])))
+
+        self.basic_hamiltonian.rotating_frame = None
+
+    def test_gradable_funcs(self):
+        """Tests whether all functions are gradable.
+        Checks if having a frame makes a difference, as well as 
+        all jax-compatible evaluation_modes."""
+        _wrap(grad,self.basic_hamiltonian.evaluate_generator)(1.0)
+        _wrap(grad,self.basic_hamiltonian.evaluate_rhs)(1.0,Array(np.array([0.2,0.4])))
+        
+        self.basic_hamiltonian.rotating_frame = Array(np.array([[3j, 2j], [2j, 0]]))
+
+        _wrap(grad,self.basic_hamiltonian.evaluate_generator)(1.0)
+        _wrap(grad,self.basic_hamiltonian.evaluate_rhs)(1.0,Array(np.array([0.2,0.4])))
+
+        self.basic_hamiltonian.rotating_frame = None
