--- test/dynamics/models/test_new_generator_models.py
@@ -1,651 +1,0 @@
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2020.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-# pylint: disable=invalid-name
-
-"""Tests for generator_models.py. """
-
-from jax import jit, grad
-from scipy.sparse import issparse
-import numpy as np
-import numpy.random as rand
-from scipy.linalg import expm
-from qiskit import QiskitError
-from qiskit.quantum_info.operators import Operator
-from qiskit_dynamics.models import CallableGenerator, GeneratorModel, RotatingFrame
-from qiskit_dynamics.signals import Signal, SignalList
-from qiskit_dynamics.dispatch import Array, wrap
-from ..common import QiskitDynamicsTestCase, TestJaxBase
-from .test_operator_collections import _wrap
-
-
-class TestGeneratorModel(QiskitDynamicsTestCase):
-    """Tests for GeneratorModel."""
-
-    def setUp(self):
-        self.X = Array(Operator.from_label("X").data)
-        self.Y = Array(Operator.from_label("Y").data)
-        self.Z = Array(Operator.from_label("Z").data)
-
-        # define a basic model
-        w = 2.0
-        r = 0.5
-        operators = [-1j * 2 * np.pi * self.Z / 2, -1j * 2 * np.pi * r * self.X / 2]
-        signals = [w, Signal(1.0, w)]
-
-        self.w = 2
-        self.r = r
-        self.basic_model = GeneratorModel(operators=operators, signals=signals)
-
-    def test_diag_frame_operator_basic_model(self):
-        """Test setting a diagonal frame operator for the internally
-        set up basic model.
-        """
-
-        self._basic_frame_evaluate_generator_test(Array([1j, -1j]), 1.123)
-        self._basic_frame_evaluate_generator_test(Array([1j, -1j]), np.pi)
-
-    def test_non_diag_frame_operator_basic_model(self):
-        """Test setting a non-diagonal frame operator for the internally
-        set up basic model.
-        """
-        self._basic_frame_evaluate_generator_test(-1j * (self.Y + self.Z), 1.123)
-        self._basic_frame_evaluate_generator_test(-1j * (self.Y - self.Z), np.pi)
-
-    def _basic_frame_evaluate_generator_test(self, frame_operator, t):
-        """Routine for testing setting of valid frame operators using the
-        basic_model.
-        """
-
-        self.basic_model.rotating_frame = frame_operator
-
-        # convert to 2d array
-        if isinstance(frame_operator, Operator):
-            frame_operator = Array(frame_operator.data)
-        if isinstance(frame_operator, Array) and frame_operator.ndim == 1:
-            frame_operator = np.diag(frame_operator)
-
-        value = self.basic_model(t)
-
-        i2pi = -1j * 2 * np.pi
-
-        U = expm(-np.array(frame_operator) * t)
-
-        # drive coefficient
-        d_coeff = self.r * np.cos(2 * np.pi * self.w * t)
-
-        # manually evaluate frame
-        expected = (
-            i2pi * self.w * U @ self.Z.data @ U.conj().transpose() / 2
-            + d_coeff * i2pi * U @ self.X.data @ U.conj().transpose() / 2
-            - frame_operator
-        )
-
-        self.assertAllClose(value, expected)
-
-    def test_evaluate_no_frame_basic_model(self):
-        """Test evaluation without a frame in the basic model."""
-
-        t = 3.21412
-        value = self.basic_model(t)
-        i2pi = -1j * 2 * np.pi
-        d_coeff = self.r * np.cos(2 * np.pi * self.w * t)
-        expected = i2pi * self.w * self.Z.data / 2 + i2pi * d_coeff * self.X.data / 2
-
-        self.assertAllClose(value, expected)
-
-    def test_evaluat_generator_in_frame_basis_basic_model(self):
-        """Test generator evaluation in frame basis in the basic_model."""
-
-        frame_op = -1j * (self.X + 0.2 * self.Y + 0.1 * self.Z).data
-
-        # enter the frame given by the -1j * X
-        self.basic_model.rotating_frame = frame_op
-
-        # get the frame basis that is used in model
-        _, U = np.linalg.eigh(1j * frame_op)
-
-        t = 3.21412
-        value = self.basic_model(t, in_frame_basis=True)
-
-        # compose the frame basis transformation with the exponential
-        # frame rotation (this will be multiplied on the right)
-        U = expm(np.array(frame_op) * t) @ U
-        Uadj = U.conj().transpose()
-
-        i2pi = -1j * 2 * np.pi
-        d_coeff = self.r * np.cos(2 * np.pi * self.w * t)
-        expected = (
-            Uadj
-            @ (i2pi * self.w * self.Z.data / 2 + i2pi * d_coeff * self.X.data / 2 - frame_op)
-            @ U
-        )
-
-        self.assertAllClose(value, expected)
-
-    def test_evaluate_generator_pseudorandom(self):
-        """Test evaluate with pseudorandom inputs."""
-
-        rng = np.random.default_rng(30493)
-        num_terms = 3
-        dim = 5
-        b = 1.0  # bound on size of random terms
-        rand_op = rng.uniform(low=-b, high=b, size=(dim, dim)) + 1j * rng.uniform(
-            low=-b, high=b, size=(dim, dim)
-        )
-        frame_op = Array(rand_op - rand_op.conj().transpose())
-        randoperators = rng.uniform(low=-b, high=b, size=(num_terms, dim, dim)) + 1j * rng.uniform(
-            low=-b, high=b, size=(num_terms, dim, dim)
-        )
-
-        rand_coeffs = Array(
-            rng.uniform(low=-b, high=b, size=(num_terms))
-            + 1j * rng.uniform(low=-b, high=b, size=(num_terms))
-        )
-        rand_carriers = Array(rng.uniform(low=-b, high=b, size=(num_terms)))
-        rand_phases = Array(rng.uniform(low=-b, high=b, size=(num_terms)))
-
-        self._test_evaluate_generator(
-            frame_op, randoperators, rand_coeffs, rand_carriers, rand_phases
-        )
-
-        rng = np.random.default_rng(94818)
-        num_terms = 5
-        dim = 10
-        b = 1.0  # bound on size of random terms
-        rand_op = rng.uniform(low=-b, high=b, size=(dim, dim)) + 1j * rng.uniform(
-            low=-b, high=b, size=(dim, dim)
-        )
-        frame_op = Array(rand_op - rand_op.conj().transpose())
-        randoperators = Array(
-            rng.uniform(low=-b, high=b, size=(num_terms, dim, dim))
-            + 1j * rng.uniform(low=-b, high=b, size=(num_terms, dim, dim))
-        )
-
-        rand_coeffs = Array(
-            rng.uniform(low=-b, high=b, size=(num_terms))
-            + 1j * rng.uniform(low=-b, high=b, size=(num_terms))
-        )
-        rand_carriers = Array(rng.uniform(low=-b, high=b, size=(num_terms)))
-        rand_phases = Array(rng.uniform(low=-b, high=b, size=(num_terms)))
-
-        self._test_evaluate_generator(
-            frame_op, randoperators, rand_coeffs, rand_carriers, rand_phases
-        )
-
-    def _test_evaluate_generator(self, frame_op, operators, coefficients, carriers, phases):
-        sig_list = []
-        for coeff, freq, phase in zip(coefficients, carriers, phases):
-
-            def get_env_func(coeff=coeff):
-                # pylint: disable=unused-argument
-                def env(t):
-                    return coeff
-
-                return env
-
-            sig_list.append(Signal(get_env_func(), freq, phase))
-        model = GeneratorModel(operators, signals=sig_list)
-        model.rotating_frame = frame_op
-
-        value = model(1.0)
-        coeffs = np.real(coefficients * np.exp(1j * 2 * np.pi * carriers * 1.0 + 1j * phases))
-
-        self.assertAllClose(
-            model.rotating_frame.operator_out_of_frame_basis(
-                model._operator_collection.evaluate_generator(coeffs)
-            ),
-            np.tensordot(coeffs, operators, axes=1) - frame_op,
-        )
-
-        expected = (
-            expm(-np.array(frame_op))
-            @ np.tensordot(coeffs, operators, axes=1)
-            @ expm(np.array(frame_op))
-            - frame_op
-        )
-
-        self.assertAllClose(value, expected)
-        if value.backend != "jax":
-            model.set_evaluation_mode("sparse")
-            state = np.arange(operators.shape[-1]*4).reshape(operators.shape[-1],4)
-            value = model(1.0)
-            if issparse(value):
-                value = value.toarray()
-            self.assertAllClose(value, expected)
-
-            val_with_state = model(1.0,state)
-            self.assertAllClose(val_with_state, value @ state)
-
-    def test_signal_setting(self):
-        """Test updating the signals."""
-
-        signals = [Signal(lambda t: 2 * t, 1.0), Signal(lambda t: t ** 2, 2.0)]
-        self.basic_model.signals = signals
-
-        t = 0.1
-        value = self.basic_model(t)
-        i2pi = -1j * 2 * np.pi
-        Z_coeff = (2 * t) * np.cos(2 * np.pi * 1 * t)
-        X_coeff = self.r * (t ** 2) * np.cos(2 * np.pi * 2 * t)
-        expected = i2pi * Z_coeff * self.Z.data / 2 + i2pi * X_coeff * self.X.data / 2
-        self.assertAllClose(value, expected)
-
-    def test_signal_setting_None(self):
-        """Test setting signals to None"""
-
-        self.basic_model.signals = None
-        self.assertTrue(self.basic_model.signals is None)
-
-    def test_signal_setting_incorrect_length(self):
-        """Test error being raised if signals is the wrong length."""
-
-        with self.assertRaises(QiskitError):
-            self.basic_model.signals = [1.0]
-
-    def test_evaluate_generator_analytic(self):
-        """Test for checking that with known operators that
-        the Model returns the analyticlly known values."""
-
-        test_operator_list = Array([self.X, self.Y, self.Z])
-        signals = SignalList([Signal(1, j / 3) for j in range(3)])
-        simple_model = GeneratorModel(
-            test_operator_list, drift=None, signals=signals, rotating_frame=None
-        )
-
-        res = simple_model.evaluate_generator(2)
-        self.assertAllClose(res, Array([[-0.5 + 0j, 1.0 + 0.5j], [1.0 - 0.5j, 0.5 + 0j]]))
-
-        simple_model.set_drift(np.eye(2), operator_in_frame_basis=False)
-        res = simple_model.evaluate_generator(2)
-        self.assertAllClose(res, Array([[0.5 + 0j, 1.0 + 0.5j], [1.0 - 0.5j, 1.5 + 0j]]))
-        simple_model.set_drift(None, operator_in_frame_basis=False)
-
-    def test_evaluate_generator_in_frame_basis_analytic(self):
-        """Tests evaluation in rotating frame against analytic values."""
-        test_operator_list = Array([self.X, self.Y, self.Z])
-        signals = SignalList([Signal(1, j / 3) for j in range(3)])
-        simple_model = GeneratorModel(
-            test_operator_list, drift=None, signals=signals, rotating_frame=None
-        )
-        simple_model.set_drift(np.eye(2), operator_in_frame_basis=False)
-        fop = Array([[0, 1j], [1j, 0]])
-        simple_model.rotating_frame = fop
-        res = simple_model(2, in_frame_basis=False)
-        expected = (
-            expm(np.array(-2 * fop))
-            @ (Array([[0.5 + 0j, 1.0 + 0.5j], [1.0 - 0.5j, 1.5 + 0j]]) - fop)
-            @ expm(np.array(2 * fop))
-        )
-        self.assertAllClose(res, expected)
-
-        res = simple_model(2, in_frame_basis=True)
-        expected = (
-            np.array([[1, 1], [1, -1]])
-            / np.sqrt(2)
-            @ expected
-            @ np.array([[1, 1], [1, -1]])
-            / np.sqrt(2)
-        )
-
-        self.assertAllClose(res, expected)
-
-        simple_model.set_drift(None)
-
-    def test_order_of_assigning_properties(self):
-        """Tests whether setting the frame, drift, and signals
-        in the constructor is the same as constructing without
-        and then assigning them in any order.
-        """
-
-        paulis = Array([self.X, self.Y, self.Z])
-        extra = Array(np.eye(2))
-        state = Array([0.2, 0.5])
-
-        t = 2
-
-        sarr = [Signal(1, j / 3) for j in range(3)]
-        sigvals = SignalList(sarr)(t)
-
-        farr = Array(np.array([[3j, 2j], [2j, 0]]))
-        farr2 = Array(np.array([[1j, 2], [-2, 3j]]))
-        evals, evect = np.linalg.eig(farr)
-        diafarr = np.diag(evals)
-
-        paulis_in_frame_basis = np.conjugate(np.transpose(evect)) @ paulis @ evect
-
-        ## Run checks without rotating frame for now
-        gm1 = GeneratorModel(paulis, sarr, extra)
-        gm2 = GeneratorModel(paulis, drift=extra)
-        gm2.signals = sarr
-        gm3 = GeneratorModel(paulis, drift=extra)
-        gm3.signals = SignalList(sarr)
-
-        # All should be the same, because there are no frames involved
-        t11 = gm1.evaluate_generator(t, False)
-        t12 = gm1.evaluate_generator(t, True)
-        t21 = gm2.evaluate_generator(t, False)
-        t22 = gm2.evaluate_generator(t, True)
-        t31 = gm3.evaluate_generator(t, False)
-        t32 = gm3.evaluate_generator(t, True)
-        t_analytical = Array([[0.5, 1.0 + 0.5j], [1.0 - 0.5j, 1.5]])
-
-        self.assertAllClose(t11, t12)
-        self.assertAllClose(t11, t21)
-        self.assertAllClose(t11, t22)
-        self.assertAllClose(t11, t31)
-        self.assertAllClose(t11, t32)
-        self.assertAllClose(t11, t_analytical)
-
-        # now work with a specific statevector
-        ts1 = gm1.evaluate_rhs(t, state, in_frame_basis=False)
-        ts2 = gm1.evaluate_rhs(t, state, in_frame_basis=True)
-        ts_analytical = Array([0.6 + 0.25j, 0.95 - 0.1j])
-
-        self.assertAllClose(ts1, ts2)
-        self.assertAllClose(ts1, ts_analytical)
-        self.assertAllClose(gm1(t, in_frame_basis=False) @ state, ts1)
-        self.assertAllClose(gm1(t, in_frame_basis=True) @ state, ts1)
-
-        ## Now, run checks with frame
-        # If passing a frame in the first place, operators must be in frame basis.abs
-        # Testing at the same time whether having Drift = None is an issue.
-        gm1 = GeneratorModel(paulis, signals=sarr, rotating_frame=RotatingFrame(farr))
-        gm2 = GeneratorModel(paulis, rotating_frame=farr)
-        gm2.signals = SignalList(sarr)
-        gm3 = GeneratorModel(paulis, rotating_frame=farr)
-        gm3.signals = sarr
-        # Does adding a frame after make a difference?
-        # If so, does it make a difference if we add signals or the frame first?
-        gm4 = GeneratorModel(paulis)
-        gm4.signals = sarr
-        gm4.rotating_frame = farr
-        gm5 = GeneratorModel(paulis)
-        gm5.rotating_frame = farr
-        gm5.signals = sarr
-        gm6 = GeneratorModel(paulis, signals=sarr)
-        gm6.rotating_frame = farr
-        # If we go to one frame, then transform back, does this make a difference?
-        gm7 = GeneratorModel(paulis, signals=sarr)
-        gm7.rotating_frame = farr2
-        gm7.rotating_frame = farr
-
-        t_in_frame_actual = Array(
-            np.diag(np.exp(-t * evals))
-            @ (np.tensordot(sigvals, paulis_in_frame_basis, axes=1) - diafarr)
-            @ np.diag(np.exp(t * evals))
-        )
-        tf1 = gm1.evaluate_generator(t, in_frame_basis=True)
-        tf2 = gm2.evaluate_generator(t, in_frame_basis=True)
-        tf3 = gm3.evaluate_generator(t, in_frame_basis=True)
-        tf4 = gm4.evaluate_generator(t, in_frame_basis=True)
-        tf5 = gm5.evaluate_generator(t, in_frame_basis=True)
-        tf6 = gm6.evaluate_generator(t, in_frame_basis=True)
-        tf7 = gm7.evaluate_generator(t, in_frame_basis=True)
-
-        self.assertAllClose(t_in_frame_actual, tf1)
-        self.assertAllClose(tf1, tf2)
-        self.assertAllClose(tf1, tf3)
-        self.assertAllClose(tf1, tf4)
-        self.assertAllClose(tf1, tf5)
-        self.assertAllClose(tf1, tf6)
-        self.assertAllClose(tf1, tf7)
-
-    def test_evaluate_rhs_lmult_equivalent_analytic(self):
-        """Tests whether evaluate_generator(t) @ state == evaluate_rhs(t,state)
-        for analytically known values."""
-
-        paulis = Array([self.X, self.Y, self.Z])
-        extra = Array(np.eye(2))
-
-        t = 2
-
-        farr = Array(np.array([[3j, 2j], [2j, 0]]))
-        evals, evect = np.linalg.eig(farr)
-        diafarr = np.diag(evals)
-
-        paulis_in_frame_basis = np.conjugate(np.transpose(evect)) @ paulis @ evect
-        extra_in_basis = evect.T.conj() @ extra @ evect
-
-        sarr = [Signal(1, j / 3) for j in range(3)]
-        sigvals = SignalList(sarr)(t)
-
-        t_in_frame_actual = Array(
-            np.diag(np.exp(-t * evals))
-            @ (np.tensordot(sigvals, paulis_in_frame_basis, axes=1) + extra_in_basis - diafarr)
-            @ np.diag(np.exp(t * evals))
-        )
-
-        state = Array([0.3, 0.1])
-        state_in_frame_basis = np.conjugate(np.transpose(evect)) @ state
-
-        gm1 = GeneratorModel(paulis, signals=sarr, rotating_frame=farr, drift=extra)
-        self.assertAllClose(gm1(t, in_frame_basis=True), t_in_frame_actual)
-        self.assertAllClose(
-            gm1(t, state_in_frame_basis, in_frame_basis=True),
-            t_in_frame_actual @ state_in_frame_basis,
-        )
-
-        t_not_in_frame_actual = Array(
-            expm(np.array(-t * farr))
-            @ (np.tensordot(sigvals, paulis, axes=1) + extra - farr)
-            @ expm(np.array(t * farr))
-        )
-
-        gm2 = GeneratorModel(paulis, signals=sarr, rotating_frame=farr, drift=extra)
-        self.assertAllClose(gm2(t, in_frame_basis=False), t_not_in_frame_actual)
-        self.assertAllClose(gm1(t, state, in_frame_basis=False), t_not_in_frame_actual @ state)
-
-        # now, remove the frame
-        gm1.rotating_frame = None
-        gm2.rotating_frame = None
-
-        t_expected = np.tensordot(sigvals, paulis, axes=1) + extra
-
-        state_in_frame_basis = state
-
-        self.assertAllClose(gm1.get_operators(True), gm1.get_operators(True))
-        self.assertAllClose(gm1.get_drift(True), gm1.get_drift(False))
-
-        self.assertAllClose(gm1(t, in_frame_basis=True), t_expected)
-        self.assertAllClose(gm1(t, state, in_frame_basis=True), t_expected @ state_in_frame_basis)
-        self.assertAllClose(gm2(t, in_frame_basis=False), t_expected)
-        self.assertAllClose(gm2(t, state, in_frame_basis=False), t_expected @ state_in_frame_basis)
-
-    def test_evaluate_rhs_vectorized_pseudorandom(self):
-        """Test for whether evaluating a model at m different
-        states, each with an n-length statevector, by passing
-        an (m,n) Array provides the same value as passing each
-        (n) Array individually."""
-        rand.seed(9231)
-        n = 8
-        k = 4
-        m = 2
-        t = rand.rand()
-        sig_list = SignalList([Signal(rand.rand(), rand.rand()) for j in range(k)])
-        normal_states = rand.uniform(-1, 1, (n))
-        vectorized_states = rand.uniform(-1, 1, (n, m))
-
-        operators = rand.uniform(-1, 1, (k, n, n))
-
-        gm = GeneratorModel(operators, drift=None, signals=sig_list)
-        self.assertTrue(gm.evaluate_rhs(t, normal_states).shape == (n,))
-        self.assertTrue(gm.evaluate_rhs(t, vectorized_states).shape == (n, m))
-        for i in range(m):
-            self.assertAllClose(
-                gm.evaluate_rhs(t, vectorized_states)[:, i],
-                gm.evaluate_rhs(t, vectorized_states[:, i]),
-            )
-
-        farr = rand.uniform(-1, 1, (n, n))
-        farr = farr - np.conjugate(np.transpose(farr))
-
-        gm.rotating_frame = farr
-
-        self.assertTrue(gm.evaluate_rhs(t, normal_states).shape == (n,))
-        self.assertTrue(gm.evaluate_rhs(t, vectorized_states).shape == (n, m))
-        vectorized_result = gm.evaluate_rhs(t, vectorized_states)
-        for i in range(m):
-            self.assertAllClose(
-                vectorized_result[:, i],
-                gm.evaluate_rhs(t, vectorized_states[:, i]),
-            )
-
-        vectorized_result = gm.evaluate_rhs(t, vectorized_states, in_frame_basis=True)
-
-        self.assertTrue(gm.evaluate_rhs(t, normal_states, in_frame_basis=True).shape == (n,))
-        self.assertTrue(vectorized_result.shape == (n, m))
-        for i in range(m):
-            self.assertAllClose(
-                vectorized_result[:, i],
-                gm.evaluate_rhs(t, vectorized_states[:, i], in_frame_basis=True),
-            )
-
-
-class TestGeneratorModelJax(TestGeneratorModel, TestJaxBase):
-    """Jax version of TestGeneratorModel tests.
-
-    Note: This class has no body but contains tests due to inheritance.
-    """
-    def test_jitable_funcs(self):
-        """Tests whether all functions are jitable.
-        Checks if having a frame makes a difference, as well as 
-        all jax-compatible evaluation_modes."""
-        _wrap(jit,self.basic_model.evaluate_generator)(1)
-        _wrap(jit,self.basic_model.evaluate_rhs)(1,Array(np.array([0.2,0.4])))
-
-        self.basic_model.rotating_frame = Array(np.array([[3j, 2j], [2j, 0]]))
-
-        _wrap(jit,self.basic_model.evaluate_generator)(1)
-        _wrap(jit,self.basic_model.evaluate_rhs)(1,Array(np.array([0.2,0.4])))
-
-        self.basic_model.rotating_frame = None
-    
-    def test_gradable_funcs(self):
-        """Tests whether all functions are gradable.
-        Checks if having a frame makes a difference, as well as 
-        all jax-compatible evaluation_modes."""
-        _wrap(grad,self.basic_model.evaluate_generator)(1.0)
-        _wrap(grad,self.basic_model.evaluate_rhs)(1.0,Array(np.array([0.2,0.4])))
-        
-        self.basic_model.rotating_frame = Array(np.array([[3j, 2j], [2j, 0]]))
-
-        _wrap(grad,self.basic_model.evaluate_generator)(1.0)
-        _wrap(grad,self.basic_model.evaluate_rhs)(1.0,Array(np.array([0.2,0.4])))
-
-        self.basic_model.rotating_frame = None
-
-
-        
-
-
-class TestCallableGenerator(QiskitDynamicsTestCase):
-    """Tests for CallableGenerator."""
-
-    def setUp(self):
-        self.X = Array(Operator.from_label("X").data)
-        self.Y = Array(Operator.from_label("Y").data)
-        self.Z = Array(Operator.from_label("Z").data)
-
-        # define a basic model
-        w = Array(2.0)
-        r = Array(0.5)
-        operators = [-1j * 2 * np.pi * self.Z / 2, -1j * 2 * np.pi * r * self.X / 2]
-
-        def generator(t):
-            return w * operators[0] + np.cos(2 * np.pi * w * t) * operators[1]
-
-        self.w = 2
-        self.r = r
-        self.basic_model = CallableGenerator(generator)
-        self.state = Array([2, 3])
-
-    def test_diag_frame_operator_basic_model(self):
-        """Test setting a diagonal frame operator for the internally
-        set up basic model.
-        """
-
-        self._basic_frame_evaluate_generator_test(Array([1j, -1j]), 1.123)
-        self._basic_frame_evaluate_generator_test(Array([1j, -1j]), np.pi)
-
-    def test_non_diag_frame_operator_basic_model(self):
-        """Test setting a non-diagonal frame operator for the internally
-        set up basic model.
-        """
-        self._basic_frame_evaluate_generator_test(-1j * (self.Y + self.Z), 1.123)
-        self._basic_frame_evaluate_generator_test(-1j * (self.Y - self.Z), np.pi)
-
-    def _basic_frame_evaluate_generator_test(self, frame_operator, t):
-        """Routine for testing setting of valid frame operators using the
-        basic_model.
-        """
-
-        self.basic_model.rotating_frame = frame_operator
-
-        # convert to 2d array
-        if isinstance(frame_operator, Operator):
-            frame_operator = Array(frame_operator.data)
-        if isinstance(frame_operator, Array) and frame_operator.ndim == 1:
-            frame_operator = np.diag(frame_operator)
-
-        value_without_state = self.basic_model(t)
-        value_with_state = self.basic_model(t, self.state)
-
-        i2pi = -1j * 2 * np.pi
-
-        U = expm(-np.array(frame_operator) * t)
-
-        # drive coefficient
-        d_coeff = self.r * np.cos(2 * np.pi * self.w * t)
-
-        # manually evaluate frame
-        expected = (
-            i2pi * self.w * U @ self.Z.data @ U.conj().transpose() / 2
-            + d_coeff * i2pi * U @ self.X.data @ U.conj().transpose() / 2
-            - frame_operator
-        )
-        self.assertAllClose(value_without_state, expected)
-        self.assertAllClose(value_with_state, expected @ self.state)
-
-
-class TestCallableGeneratorJax(TestCallableGenerator, TestJaxBase):
-    """Jax version of TestCallableGenerator tests.
-
-    Note: This class has no body but contains tests due to inheritance.
-    """
-
-    def test_jitable_funcs(self):
-        """Tests whether all functions are jitable.
-        Checks if having a frame makes a difference, as well as 
-        all jax-compatible evaluation_modes."""
-        _wrap(jit,self.basic_model.evaluate_generator)(1)
-        _wrap(jit,self.basic_model.evaluate_rhs)(1,Array(np.array([0.2,0.4])))
-
-        self.basic_model.rotating_frame = Array(np.array([[3j, 2j], [2j, 0]]))
-
-        _wrap(jit,self.basic_model.evaluate_generator)(1)
-        _wrap(jit,self.basic_model.evaluate_rhs)(1,Array(np.array([0.2,0.4])))
-
-        self.basic_model.rotating_frame = None
-    
-    def test_gradable_funcs(self):
-        """Tests whether all functions are gradable.
-        Checks if having a frame makes a difference, as well as 
-        all jax-compatible evaluation_modes."""
-        _wrap(grad,self.basic_model.evaluate_generator)(1.0)
-        _wrap(grad,self.basic_model.evaluate_rhs)(1.0,Array(np.array([0.2,0.4])))
-        
-        self.basic_model.rotating_frame = Array(np.array([[3j, 2j], [2j, 0]]))
-
-        _wrap(grad,self.basic_model.evaluate_generator)(1.0)
-        _wrap(grad,self.basic_model.evaluate_rhs)(1.0,Array(np.array([0.2,0.4])))
-
-        self.basic_model.rotating_frame = None
